demo: switch to explicit permissions management to access GPU
We've used implicit permissions management to access GPU, i.e. we
got required permissions via sudo calls at runtime. This is not secure
and not recommended by SDL. This change switches to explicit permissions
management.

To achieve the goal we also need to slightly change architecture and
create one more server besides nginx to handle transcoding requests.
Previously we used nginx for this purpose, but the problem is how nginx
manages users and groups. Actually it starts shell anew and we lose those
permissions which we were granted via docker run command line. To handle
that we just tart yet-another simple socat server which can execute bash
scripts for us and schedule transcoding tasks on it. Actually such
architecture is much closer to the real customer deployments.

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>