Initial commit for Media Delivery Software Stack
This project provides samples to demonstrate the use of Intel GPU
in simplified real-life scenarios involving media streams delivery.
It leverages a Software Stack which consists of the following
ingredients:
* Intel Media SDK
* Intel Media Driver
* FFmpeg w/ enabled ffmpeg-qsv plugins

Provided samples focus on the key aspects of the proper Intel software
setup and integration with other popular tools you will likely use in
your final product. We try to do our best to provide a configuration
which will demonstrate the best quality and performance for Intel GPU
media stack. Key topcis we are covering:
* Samples which demonstrate operations typical for Content Delivery
  Network (CDN) applications such as Video On Demand (VOD) streaming
  under Nginx server
* Reference command lines (for ffmpeg-qsv and mediasdk native samples)
  tuned for the optimal quality and performance (for the showcasing
  scenario)
* Quality and Performance measuring infrastructure for data
  collection
* Intel GPU Performance monitoring

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>