Switch to selfbuild for default builds
To facilitate regular project development (which means to be
able to update key components on development needs) we need
a more flexible way to fetch our dependencies. dgpu-docs.intel.com
provides packages of the older versions which we can't use
for project development. This patch switches the project
to just selfbuild required components.

Dockerfile to build against dgpu-docs.intel.com is remained
there, but might fail to build or built stack might fail to
work if packages are too old. We might still need this Dockerfile
to be able to build against custom set of packages via self
configured apt repository or if we will need to support a
release against packages exposed on dgpu-docs.intel.com.

This change also clarifies latest direction for DG1 - use
upstream stack for that platform instead of customized production
kernel.

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>