vmaf: install python modules to pip default location
Ubuntu 22.04 adjusted how python modules are installed and our
current method stopped working (under root they install to
dist-packages vs. site-packages as was on Ubuntu 20.04).

Since by default pip installs to /usr/local, it does not make
sense to continue installing in yet-another custom location
considering this is not well tunable with pip. So, with this
commit we just rely on pip default behavior.

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>