Make changes to vsync mode in the ffmpeg scripts
We cannot use `-fps_mode passthrough` combined with a fixed frame rate via `-r`. That will fail as shown:

Error reported by ffmpeg will be: 
`sh
One of -r/-fpsmax was specified together a non-CFR -vsync/-fps_mode. This is contradictory.
```
To correct this, either:

1. Either set  `-fps_mode` to either `auto` or `cfr` or omit the option entirely when `-r` is set/configured;
2. Drop the `-r` option entirely and set the output framerate via ffmpeg's `fps` filter (not recommended for QSV and VAAPI) or via `vpp_qsv`'s `fps` argument (recommended) if `-fps_mode passthrough` has to be configured/set.

For now, I'll update the documentation with  `-fps_mode auto` toggled on so that the existing command-lines don't break.