doc: fix rst-lint reported issues
This patch fixes all rst-lint reported issues with level >= warning:
  WARNING ./doc/quality.rst:205 Blank line required after table.
  WARNING ./doc/quality-lowdelay.rst:180 Blank line required after table.
  WARNING ./doc/benchmarking.rst:2 Duplicate explicit target name: "performance".
  WARNING ./doc/benchmarking.rst:2 Duplicate explicit target name: "quality".
  WARNING ./doc/benchmarking.rst:2 Duplicate explicit target name: "performance".
  WARNING ./doc/benchmarking.rst:2 Duplicate explicit target name: "quality".
  WARNING ./doc/virtualization.rst:455 Bullet list ends without a blank line; unexpected unindent.
  WARNING ./doc/benchmarks/intel-data-center-gpu-flex-series/intel-data-center-gpu-flex-series.rst:2 Duplicate explicit target name: "22ww33_419.38_ubuntu514".

It also fixes some, but not all issues with level == info. Specifically,
"duplicate implicit target name" issues are not fixed if reported for
duplicated sub-headers.

It occurs that anonymous links should be given with `__` (2 underscores) in rst.
There is a reason however to leave non-anonymous links (`_`, i.e. 1 underscore)
where possible, which is everywhere except when we get an issue from rst-lint
that anonymous link is duplicated. Such cases got fixed in this change.

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>